requirements.txt
requests
psutil
argparse
pyusb